requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.5
pandas==2.2.3
openpyxl==3.1.2
Pillow==10.4.0
python-dotenv==1.0.0
Authlib==1.3.1
requests==2.32.3
APScheduler==3.10.4
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
reportlab==4.2.5
# Prod-only servers; skip on Windows to avoid build issues
gunicorn==21.2.0; sys_platform != 'win32'
gevent==24.2.1; sys_platform != 'win32'
twilio>=9,<10
webdriver-manager==4.0.2
selenium==4.38.0

# Optional Postgres driver (for DATABASE_URL). Safe on Linux deploy targets.
psycopg2-binary==2.9.9
